requirements.txt
pandas>=1.0.3
rpy2>=3.3.2
gitpython
